+++ processing/process_election_results.py
@@ -5,14 +5,16 @@ This script processes election results data for geospatial visualization and upl
 as the primary data backend. It handles geospatial data, defines calculated fields, and prepares
 comprehensive election data for web consumption and API integration.
 
+Now refactored to use the spatial_utils module for all spatial operations.
+
 Key Functionality:
 1. Field Definitions:
    - Registers calculated fields (e.g., percentages, counts, ratios) with explanations and formulas.
    - Maintains a field registry for documentation and validation.
 
-2. Geospatial Processing:
-   - Validates and reprojects geospatial data to WGS84 (standard for mapping).
-   - Optimizes properties for efficient rendering in web maps and PostGIS queries.
+2. Data Processing:
+   - Adds analytical fields and competition metrics to election data.
+   - Calculates vote margins, turnout rates, and demographic correlations.
 
 3. Data Validation:
    - Validates the completeness of fields in the geospatial dataset.
@@ -45,6 +47,7 @@ Example:
 Dependencies:
 - geopandas, pandas, numpy, matplotlib, loguru, pathlib, and other standard Python libraries.
 - Supabase integration (optional): sqlalchemy, psycopg2-binary for database uploads.
+- Spatial operations: spatial_utils module for all geometric processing.
 """
 
 from typing import Any, Callable, Dict, List, Optional
@@ -62,37 +65,17 @@ from ops.field_registry import (
     generate_layer_explanations,
 )
 
-# Import Supabase integration
-try:
-    from ops.repositories import SpatialQueryManager
-    from ops.supabase_integration import SupabaseDatabase, SupabaseUploader
-
-    logger.debug("✅ Imported Supabase integration with improved design patterns")
-    SUPABASE_AVAILABLE = True
-except ImportError as e:
-    logger.debug(f"📊 Supabase integration not available: {e}")
-    logger.debug("   Install with: pip install sqlalchemy psycopg2-binary")
-    SUPABASE_AVAILABLE = False
-
-    # Fallback classes for when Supabase integration is not available
-    class SupabaseUploader:
-        def __init__(self, config):
-            pass
-
-        def upload_geodataframe(self, *args, **kwargs):
-            return False
-
-    class SupabaseDatabase:
-        def __init__(self, config):
-            pass
-
-    class SpatialQueryManager:
-        def __init__(self, db):
-            pass
-
-        def get_sample_records(self, *args, **kwargs):
-            return []
-
+# Import spatial operations from spatial_utils module
+from spatial_utils import (
+    validate_and_reproject_to_wgs84,
+    optimize_geojson_properties,
+    clean_numeric,
+    consolidate_split_precincts,
+    SUPABASE_AVAILABLE,
+    SupabaseUploader,
+    SupabaseDatabase,
+    SpatialQueryManager
+)
 
 # Global registry instance
 FIELD_REGISTRY = FieldRegistry()
@@ -241,396 +224,6 @@ def detect_contribution_columns(gdf: gpd.GeoDataFrame) -> List[str]:
     ]
     logger.debug(f"  📊 Detected contribution columns: {contribution_cols}")
     return contribution_cols
-
-
-def consolidate_split_precincts(gdf: gpd.GeoDataFrame, precinct_col: str) -> gpd.GeoDataFrame:
-    """
-    Consolidate split precincts (e.g., 2801a, 2801b, 2801c) into single features.
-    FIXED to preserve county rollup data properly.
-
-    Args:
-        gdf: GeoDataFrame with precinct data
-        precinct_col: Name of the precinct column
-
-    Returns:
-        GeoDataFrame with consolidated precincts
-    """
-    logger.debug(f"🔄 Consolidating split precincts in column '{precinct_col}':")
-
-    # Create a copy to work with
-    gdf_work = gdf.copy()
-
-    # Convert ALL numeric columns to proper numeric types BEFORE processing
-    logger.debug("  🔧 Converting columns to proper data types...")
-
-    # Identify boolean columns first to exclude them from numeric conversion
-    boolean_cols = [
-        "is_pps_precinct",
-        "has_election_results",
-        "has_voter_registration",
-        "is_summary",
-        "is_complete_record",
-        "is_county_rollup",
-    ]
-
-    # Identify categorical columns that should NOT be converted to numeric
-    categorical_cols = [
-        "political_lean",
-        "competitiveness",
-        "leading_candidate",
-        "second_candidate",
-        "record_type",
-        "turnout_quartile",
-        "margin_category",
-        "precinct_size_category",
-    ]
-
-    # Identify ALL columns that should be numeric and convert them (excluding boolean and categorical columns)
-    numeric_conversion_cols = []
-    for col in gdf_work.columns:
-        if col in ["geometry", precinct_col, "base_precinct"] + boolean_cols + categorical_cols:
-            continue
-        # Check if this looks like a numeric column based on content
-        sample_values = gdf_work[col].dropna().head(10)
-        if len(sample_values) > 0:
-            # Try to convert sample to see if it's numeric
-            try:
-                converted = pd.to_numeric(sample_values, errors="coerce")
-                # Only include if the conversion actually worked (not all NaN)
-                if not converted.isna().all():
-                    numeric_conversion_cols.append(col)
-            except Exception as e:
-                logger.error(f"Error converting column {col} to numeric: {e}")
-
-    # Convert identified numeric columns (excluding booleans and categoricals)
-    for col in numeric_conversion_cols:
-        gdf_work[col] = pd.to_numeric(gdf_work[col], errors="coerce").fillna(0)
-
-    # Handle boolean columns separately - convert to proper boolean type
-    for col in boolean_cols:
-        if col in gdf_work.columns:
-            gdf_work[col] = gdf_work[col].astype(str).str.lower().isin(["true", "1", "yes"])
-
-    logger.debug(f"  📊 Converted {len(numeric_conversion_cols)} columns to numeric")
-    logger.debug(
-        f"  📊 Converted {sum(1 for col in boolean_cols if col in gdf_work.columns)} columns to boolean"
-    )
-
-    # Extract base precinct numbers (remove a,b,c suffixes)
-    gdf_work["base_precinct"] = (
-        gdf_work[precinct_col].astype(str).str.replace(r"[a-zA-Z]+$", "", regex=True).str.strip()
-    )
-
-    # Count how many precincts have splits
-    precinct_counts = gdf_work["base_precinct"].value_counts()
-    split_precincts = precinct_counts[precinct_counts > 1]
-
-    logger.debug(f"  📊 Found {len(split_precincts)} precincts with splits:")
-    for base, count in split_precincts.head(10).items():
-        logger.debug(f"    - Precinct {base}: {count} parts")
-    if len(split_precincts) > 10:
-        logger.debug(f"    ... and {len(split_precincts) - 10} more")
-
-    # Identify ALL numeric columns to take first value during consolidation
-    numeric_cols = []
-    for col in gdf_work.columns:
-        if col in ["geometry", precinct_col, "base_precinct"]:
-            continue
-        # Include ANY column that starts with numeric prefixes OR is explicitly numeric
-        if (
-            col.startswith(
-                (
-                    "votes_",
-                    "TOTAL",
-                    "DEM",
-                    "REP",
-                    "NAV",
-                    "OTH",
-                    "CON",
-                    "IND",
-                    "LBT",
-                    "NLB",
-                    "PGP",
-                    "PRO",
-                    "WFP",
-                    "WTP",
-                )
-            )
-            or col in ["vote_margin", "total_voters"]
-            or gdf_work[col].dtype in ["int64", "float64"]
-        ):
-            numeric_cols.append(col)
-
-    # Identify percentage/rate columns - FIXED for new percentage scale (0-100)
-    percentage_cols = []
-    for col in gdf_work.columns:
-        if col.startswith(("vote_pct_", "reg_pct_")) or col in [
-            "turnout_rate",
-            "dem_advantage",
-            "major_party_pct",
-            "margin_pct",
-            "engagement_score",
-        ]:
-            percentage_cols.append(col)
-
-    logger.debug(
-        f"  📊 Will take first value from {len(numeric_cols)} numeric columns during consolidation"
-    )
-    logger.debug(f"  📊 Will recalculate {len(percentage_cols)} percentage columns")
-
-    # Debug: Check vote totals BEFORE consolidation (sum of unique precincts only)
-    unique_precinct_votes = 0
-    if "votes_total" in gdf_work.columns:
-        # For accurate comparison, count only unique base precincts
-        unique_precinct_votes = gdf_work.groupby("base_precinct")["votes_total"].first().sum()
-        pre_consolidation_total = gdf_work["votes_total"].sum()  # Raw total (includes duplicates)
-        logger.debug(
-            f"  🔍 Total votes BEFORE consolidation: {pre_consolidation_total:,.0f} (raw with duplicates)"
-        )
-        logger.debug(
-            f"  🔍 Unique precinct votes: {unique_precinct_votes:,.0f} (expected after consolidation)"
-        )
-
-    # Group by base precinct and consolidate
-    consolidated_features = []
-
-    for base_precinct in gdf_work["base_precinct"].unique():
-        if pd.isna(base_precinct) or base_precinct == "":
-            continue
-
-        precinct_parts = gdf_work[gdf_work["base_precinct"] == base_precinct]
-
-        if len(precinct_parts) == 1:
-            # Single precinct, just update the precinct name to base
-            feature = precinct_parts.copy()
-            feature[precinct_col] = base_precinct
-            consolidated_features.append(feature)
-        else:
-            # Multiple parts - consolidate them
-            consolidated = precinct_parts.iloc[0:1].copy()  # Start with first part
-
-            # Update precinct identifier to base
-            consolidated[precinct_col] = base_precinct
-
-            # Take values from the first part (they should all be identical for split precincts)
-            for col in numeric_cols:
-                if col in precinct_parts.columns:
-                    # For split precincts, all parts should have identical values - take the first
-                    first_value = precinct_parts[col].iloc[0]
-                    consolidated.loc[consolidated.index[0], col] = first_value
-
-                    # Debug: Verify all parts have same values for vote columns
-                    if col.startswith("votes_") and first_value > 0:
-                        all_values = precinct_parts[col].tolist()
-                        if len(set(all_values)) == 1:
-                            logger.debug(
-                                f"    🔍 {base_precinct} {col}: {first_value} (verified identical across {len(all_values)} parts)"
-                            )
-                        else:
-                            logger.debug(
-                                f"    ⚠️  {base_precinct} {col}: Values differ across parts: {all_values} - taking first: {first_value}"
-                            )
-
-            # Handle boolean and categorical columns properly
-            categorical_cols = [
-                "political_lean",
-                "competitiveness",
-                "leading_candidate",
-                "record_type",
-                "turnout_quartile",
-                "margin_category",
-                "precinct_size_category",
-            ]
-
-            # For boolean columns, use logical OR (if ANY part is True, consolidated should be True)
-            for col in boolean_cols:
-                if col in precinct_parts.columns:
-                    # Convert to boolean and take logical OR
-                    bool_values = precinct_parts[col].astype(
-                        bool
-                    )  # Already converted to bool above
-                    consolidated_value = bool_values.any()  # True if ANY part is True
-                    consolidated.loc[consolidated.index[0], col] = (
-                        consolidated_value  # Use .loc for proper assignment
-                    )
-
-                    if col == "is_pps_precinct" and consolidated_value:
-                        logger.debug(
-                            f"    🔍 {base_precinct} {col}: {precinct_parts[col].tolist()} → {consolidated_value}"
-                        )
-
-            # For categorical columns, use the first value (should be identical for split precincts)
-            for col in categorical_cols:
-                if col in precinct_parts.columns:
-                    # Take the first value (should be same across all parts for split precincts)
-                    first_value = precinct_parts[col].iloc[0]
-                    consolidated.loc[consolidated.index[0], col] = first_value
-
-            # Dissolve geometries (combine all parts into one shape) - FIXED for edge artifacts
-            try:
-                # STEP 1: Clean geometries first to avoid edge artifacts
-                cleaned_geoms = []
-                for idx, geom in enumerate(precinct_parts.geometry):
-                    if geom is not None and geom.is_valid:
-                        # Apply tiny buffer to clean up potential edge issues
-                        cleaned_geom = geom.buffer(0.0000001).buffer(
-                            -0.0000001
-                        )  # Tiny buffer + negative buffer
-                        if cleaned_geom.is_valid and not cleaned_geom.is_empty:
-                            cleaned_geoms.append(cleaned_geom)
-                        else:
-                            logger.warning(
-                                f"      ⚠️ Geometry {idx} invalid after cleaning, using original"
-                            )
-                            cleaned_geoms.append(geom)
-                    elif geom is not None:
-                        # Try to fix invalid geometry with standard buffer
-                        try:
-                            fixed_geom = geom.buffer(0)
-                            if fixed_geom.is_valid and not fixed_geom.is_empty:
-                                cleaned_geoms.append(fixed_geom)
-                            else:
-                                logger.warning(f"      ⚠️ Could not fix geometry {idx}, skipping")
-                        except Exception:
-                            logger.warning(f"      ⚠️ Could not process geometry {idx}, skipping")
-
-                # STEP 2: Dissolve using cleaned geometries
-                if cleaned_geoms:
-                    # Use unary_union on cleaned geometries
-                    dissolved_geom = gpd.GeoSeries(cleaned_geoms).unary_union
-
-                    # STEP 3: Final validation and cleanup
-                    if dissolved_geom.is_valid:
-                        # Apply final tiny buffer to ensure clean edges
-                        final_geom = dissolved_geom.buffer(0.0000001).buffer(-0.0000001)
-                        if final_geom.is_valid and not final_geom.is_empty:
-                            consolidated.loc[consolidated.index[0], "geometry"] = final_geom
-                        else:
-                            consolidated.loc[consolidated.index[0], "geometry"] = dissolved_geom
-                    else:
-                        # Try to fix with standard buffer
-                        fixed_geom = dissolved_geom.buffer(0)
-                        if fixed_geom.is_valid:
-                            consolidated.loc[consolidated.index[0], "geometry"] = fixed_geom
-                        else:
-                            logger.debug(
-                                f"    ⚠️ Could not create valid dissolved geometry for {base_precinct}, using first part"
-                            )
-                            consolidated.loc[consolidated.index[0], "geometry"] = (
-                                precinct_parts.geometry.iloc[0]
-                            )
-                else:
-                    logger.warning(
-                        f"    ⚠️ No valid geometries to dissolve for {base_precinct}, using first part"
-                    )
-                    consolidated.loc[consolidated.index[0], "geometry"] = (
-                        precinct_parts.geometry.iloc[0]
-                    )
-
-            except Exception as e:
-                logger.warning(f"    ⚠️ Error dissolving geometry for precinct {base_precinct}: {e}")
-                # Use the first geometry as fallback
-                consolidated.loc[consolidated.index[0], "geometry"] = precinct_parts.geometry.iloc[
-                    0
-                ]
-
-            # Keep other fields from first part for remaining columns
-            # Note: percentage columns will be recalculated later based on new totals
-
-            consolidated_features.append(consolidated)
-
-    # Combine all consolidated features
-    if consolidated_features:
-        gdf_consolidated = pd.concat(consolidated_features, ignore_index=True)
-
-        # Debug: Check vote totals AFTER consolidation
-        if "votes_total" in gdf_consolidated.columns:
-            post_consolidation_total = gdf_consolidated["votes_total"].sum()
-            logger.debug(f"  🔍 Total votes AFTER consolidation: {post_consolidation_total:,.0f}")
-            if unique_precinct_votes > 0:
-                retention_rate = (post_consolidation_total / unique_precinct_votes) * 100
-                logger.debug(f"  📊 Vote retention rate: {retention_rate:.1f}% (should be 100%)")
-            else:
-                logger.debug("  📊 Vote total preserved correctly")
-
-        # Recalculate percentage columns based on new totals - FIXED for percentage scale
-        logger.debug("  🔄 Recalculating percentage columns...")
-        for col in percentage_cols:
-            if col in gdf_consolidated.columns:
-                if col.startswith("vote_pct_"):
-                    # Find the corresponding count column
-                    count_col = col.replace("vote_pct_", "votes_")
-                    if (
-                        count_col in gdf_consolidated.columns
-                        and "votes_total" in gdf_consolidated.columns
-                    ):
-                        # Convert to numeric and recalculate percentages (0-100 scale)
-                        count_values = pd.to_numeric(
-                            gdf_consolidated[count_col], errors="coerce"
-                        ).fillna(0)
-                        total_values = pd.to_numeric(
-                            gdf_consolidated["votes_total"], errors="coerce"
-                        ).fillna(0)
-                        gdf_consolidated[col] = np.where(
-                            total_values > 0,
-                            (count_values / total_values) * 100,  # Scale to 0-100
-                            0,
-                        )
-                elif (
-                    col == "turnout_rate"
-                    and "votes_total" in gdf_consolidated.columns
-                    and "TOTAL" in gdf_consolidated.columns
-                ):
-                    # Recalculate turnout rate (0-100 scale)
-                    vote_values = pd.to_numeric(
-                        gdf_consolidated["votes_total"], errors="coerce"
-                    ).fillna(0)
-                    total_values = pd.to_numeric(gdf_consolidated["TOTAL"], errors="coerce").fillna(
-                        0
-                    )
-                    gdf_consolidated[col] = np.where(
-                        total_values > 0,
-                        (vote_values / total_values) * 100,  # Scale to 0-100
-                        0,
-                    )
-                elif col == "dem_advantage":
-                    # Recalculate dem_advantage (already on 0-100 scale)
-                    if (
-                        "reg_pct_dem" in gdf_consolidated.columns
-                        and "reg_pct_rep" in gdf_consolidated.columns
-                    ):
-                        dem_values = pd.to_numeric(
-                            gdf_consolidated["reg_pct_dem"], errors="coerce"
-                        ).fillna(0)
-                        rep_values = pd.to_numeric(
-                            gdf_consolidated["reg_pct_rep"], errors="coerce"
-                        ).fillna(0)
-                        gdf_consolidated[col] = dem_values - rep_values
-                elif col == "major_party_pct":
-                    # Recalculate major_party_pct (already on 0-100 scale)
-                    if (
-                        "reg_pct_dem" in gdf_consolidated.columns
-                        and "reg_pct_rep" in gdf_consolidated.columns
-                    ):
-                        dem_values = pd.to_numeric(
-                            gdf_consolidated["reg_pct_dem"], errors="coerce"
-                        ).fillna(0)
-                        rep_values = pd.to_numeric(
-                            gdf_consolidated["reg_pct_rep"], errors="coerce"
-                        ).fillna(0)
-                        gdf_consolidated[col] = dem_values + rep_values
-
-        logger.debug(
-            f"  ✅ Consolidated {len(gdf_work)} features into {len(gdf_consolidated)} features"
-        )
-        logger.debug(
-            f"  ✅ Eliminated {len(gdf_work) - len(gdf_consolidated)} duplicate/split features"
-        )
-
-        return gdf_consolidated
-    else:
-        logger.warning("  ⚠️ Warning: No features to consolidate")
-        return gdf_work
 
 
 def add_analytical_fields(df: pd.DataFrame) -> pd.DataFrame:
@@ -1288,438 +881,6 @@ def add_analytical_fields(df: pd.DataFrame) -> pd.DataFrame:
     )
 
     return df_analysis
-
-
-def clean_numeric(series: pd.Series, is_percent: bool = False) -> pd.Series:
-    """
-    Cleans a pandas Series to numeric type, handling commas and percent signs.
-    FIXED for new percentage data scale (already 0-100, don't divide by 100 again).
-
-    Args:
-        series: The pandas Series to clean.
-        is_percent: If True, data is already in percentage format (0-100), don't convert
-
-    Returns:
-        A pandas Series with numeric data.
-    """
-    s = (
-        series.astype(str)
-        .str.replace(",", "", regex=False)
-        .str.replace("%", "", regex=False)
-        .str.strip()
-    )
-    vals = pd.to_numeric(s, errors="coerce")
-    # Don't divide by 100 - our new data is already in percentage format
-    return vals
-
-
-def validate_and_reproject_to_wgs84(
-    gdf: gpd.GeoDataFrame, config: Config, source_description: str = "GeoDataFrame"
-) -> gpd.GeoDataFrame:
-    """
-    Validates and reprojects a GeoDataFrame to WGS84 (EPSG:4326) if needed.
-
-    Args:
-        gdf: Input GeoDataFrame
-        config: Configuration instance
-        source_description: Description for logging
-
-    Returns:
-        GeoDataFrame in WGS84 coordinate system
-    """
-    logger.debug(f"🗺️ Validating and reprojecting {source_description}:")
-
-    # Check original CRS
-    original_crs = gdf.crs
-    logger.debug(f"  📍 Original CRS: {original_crs}")
-
-    # Get CRS settings from config
-    input_crs = config.get_system_setting("input_crs")
-    output_crs = config.get_system_setting("output_crs")
-
-    # Handle missing CRS
-    if original_crs is None:
-        logger.warning("  ⚠️ No CRS specified in data")
-
-        # Try to detect coordinate system from sample coordinates
-        if not gdf.empty and "geometry" in gdf.columns:
-            sample_geom = gdf.geometry.dropna().iloc[0] if len(gdf.geometry.dropna()) > 0 else None
-            if sample_geom is not None:
-                # Get first coordinate pair
-                coords = None
-                if hasattr(sample_geom, "exterior"):  # Polygon
-                    coords = list(sample_geom.exterior.coords)[0]
-                elif hasattr(sample_geom, "coords"):  # Point or LineString
-                    coords = list(sample_geom.coords)[0]
-
-                if coords:
-                    x, y = coords[0], coords[1]
-                    logger.debug(f"  🔍 Sample coordinates: x={x:.2f}, y={y:.2f}")
-
-                    # Check if coordinates look like configured input CRS
-                    if input_crs == "EPSG:2913" and abs(x) > 1000000 and abs(y) > 1000000:
-                        logger.debug(f"  🎯 Coordinates appear to be {input_crs}")
-                        gdf = gdf.set_crs(input_crs, allow_override=True)
-                    # Check if coordinates look like WGS84 (longitude/latitude)
-                    elif -180 <= x <= 180 and -90 <= y <= 90:
-                        logger.debug(f"  🎯 Coordinates appear to be {output_crs}")
-                        gdf = gdf.set_crs(output_crs, allow_override=True)
-                    else:
-                        logger.warning(f"  ❓ Unknown coordinate system, assuming {input_crs}")
-                        gdf = gdf.set_crs(input_crs, allow_override=True)
-                else:
-                    logger.warning(
-                        f"  ❓ Could not extract sample coordinates, assuming {output_crs}"
-                    )
-                    gdf = gdf.set_crs(output_crs, allow_override=True)
-            else:
-                logger.warning(f"  ❓ No valid geometry found, assuming {output_crs}")
-                gdf = gdf.set_crs(output_crs, allow_override=True)
-
-    # Reproject to output CRS if needed
-    current_crs = gdf.crs
-    if current_crs is not None:
-        try:
-            current_epsg = current_crs.to_epsg()
-            target_epsg = int(output_crs.split(":")[1])
-            if current_epsg != target_epsg:
-                logger.debug(f"  🔄 Reprojecting from EPSG:{current_epsg} to {output_crs}")
-                gdf_reprojected = gdf.to_crs(output_crs)
-
-                # Validate reprojection worked
-                if not gdf_reprojected.empty and "geometry" in gdf_reprojected.columns:
-                    sample_geom = (
-                        gdf_reprojected.geometry.dropna().iloc[0]
-                        if len(gdf_reprojected.geometry.dropna()) > 0
-                        else None
-                    )
-                    if sample_geom is not None:
-                        coords = None
-                        if hasattr(sample_geom, "exterior"):  # Polygon
-                            coords = list(sample_geom.exterior.coords)[0]
-                        elif hasattr(sample_geom, "coords"):  # Point or LineString
-                            coords = list(sample_geom.coords)[0]
-
-                        if coords:
-                            x, y = coords[0], coords[1]
-                            logger.debug(f"  ✓ Reprojected coordinates: lon={x:.6f}, lat={y:.6f}")
-
-                            # Validate coordinates are in valid WGS84 range
-                            if -180 <= x <= 180 and -90 <= y <= 90:
-                                logger.debug("  ✓ Coordinates are valid WGS84")
-                            else:
-                                logger.warning(f"  ⚠️ Coordinates may be invalid: lon={x}, lat={y}")
-                        else:
-                            logger.warning("  ⚠️ Could not validate reprojected coordinates")
-
-                gdf = gdf_reprojected
-            else:
-                logger.debug(f"  ✓ Already in {output_crs}")
-        except Exception as e:
-            logger.error(f"  ❌ Error during reprojection: {e}")
-            logger.debug(f"  🔧 Attempting to set CRS as {output_crs}")
-            gdf = gdf.set_crs(output_crs, allow_override=True)
-
-    # Final validation
-    if gdf.crs is not None:
-        try:
-            final_epsg = gdf.crs.to_epsg()
-            logger.debug(f"  ✅ Final CRS: EPSG:{final_epsg}")
-        except Exception:
-            logger.error(f"  ❌ Final CRS: {gdf.crs}")
-    else:
-        logger.warning("  ⚠️ Warning: Final CRS is None")
-
-    # Validate geometry
-    valid_geom_count = gdf.geometry.notna().sum()
-    total_count = len(gdf)
-    logger.debug(
-        f"  📊 Geometry validation: {valid_geom_count}/{total_count} features have valid geometry"
-    )
-
-    return gdf
-
-
-def optimize_geojson_properties(gdf: gpd.GeoDataFrame, config: Config) -> gpd.GeoDataFrame:
-    """
-    Optimizes GeoDataFrame properties for web display and vector tile generation.
-    Uses field registry and duck-typing instead of hardcoded column names.
-
-    Args:
-        gdf: Input GeoDataFrame
-        config: Configuration instance
-
-    Returns:
-        GeoDataFrame with optimized properties
-    """
-    logger.debug("🔧 Optimizing properties for web display using dynamic field detection:")
-
-    # Create a copy to avoid modifying original
-    gdf_optimized = gdf.copy()
-
-    # Get precision settings from config
-    prop_precision = config.get_system_setting("property_precision")
-
-    # Clean up property names and values for web consumption
-    columns_to_clean = gdf_optimized.columns.tolist()
-    if "geometry" in columns_to_clean:
-        columns_to_clean.remove("geometry")
-
-    # Track optimization stats and collect optimized columns
-    optimized_counts = {
-        "boolean": 0,
-        "count": 0,
-        "percentage": 0,
-        "categorical": 0,
-        "identifier": 0,
-        "unknown": 0,
-    }
-
-    # Collect all optimized columns to update at once
-    optimized_data = {}
-
-    for col in columns_to_clean:
-        if col in gdf_optimized.columns:
-            series = gdf_optimized[col]
-
-            # Get field info from registry if available
-            field_def = FIELD_REGISTRY._fields.get(col)
-            field_type = field_def.field_type if field_def else None
-
-            # 1. BOOLEAN FIELDS - Use registry + duck typing
-            if (
-                field_type == "boolean"
-                or col.startswith(("is_", "has_"))
-                or _is_boolean_data(series)
-            ):
-                optimized_data[col] = _optimize_boolean_field(series)
-                optimized_counts["boolean"] += 1
-
-            # 2. COUNT FIELDS - Use registry + pattern detection
-            elif field_type == "count" or _is_count_field(col, series):
-                optimized_data[col] = _optimize_count_field(series)
-                optimized_counts["count"] += 1
-
-            # 3. PERCENTAGE FIELDS - Use registry + pattern detection
-            elif field_type == "percentage" or _is_percentage_field(col, series):
-                optimized_data[col] = _optimize_percentage_field(series, prop_precision)
-                optimized_counts["percentage"] += 1
-
-            # 4. CATEGORICAL FIELDS - Use registry + duck typing
-            elif field_type == "categorical" or _is_categorical_field(col, series):
-                optimized_data[col] = _optimize_categorical_field(col, series)
-                optimized_counts["categorical"] += 1
-
-            # 5. IDENTIFIER FIELDS - Pattern detection
-            elif _is_identifier_field(col):
-                optimized_data[col] = _optimize_identifier_field(series)
-                optimized_counts["identifier"] += 1
-
-            # 6. FALLBACK - Try to infer from data
-            else:
-                optimized_data[col] = _optimize_unknown_field(col, series, prop_precision)
-                optimized_counts["unknown"] += 1
-
-    # Update all optimized columns at once to avoid DataFrame fragmentation
-    if optimized_data:
-        gdf_optimized = gdf_optimized.assign(**optimized_data)
-
-    # Log optimization results
-    total_optimized = sum(optimized_counts.values())
-    logger.debug(f"  ✓ Optimized {total_optimized} property columns:")
-    for field_type, count in optimized_counts.items():
-        if count > 0:
-            logger.debug(f"    - {field_type}: {count} fields")
-
-    # Add web-friendly geometry validation
-    invalid_geom = gdf_optimized.geometry.isna() | (~gdf_optimized.geometry.is_valid)
-    invalid_count = invalid_geom.sum()
-
-    if invalid_count > 0:
-        logger.warning(f"  ⚠️ Found {invalid_count} invalid geometries, attempting to fix...")
-        # Try to fix invalid geometries
-        gdf_optimized.geometry = gdf_optimized.geometry.buffer(0)
-
-        # Check again
-        still_invalid = gdf_optimized.geometry.isna() | (~gdf_optimized.geometry.is_valid)
-        still_invalid_count = still_invalid.sum()
-
-        if still_invalid_count > 0:
-            logger.warning(f"  ⚠️ {still_invalid_count} geometries still invalid after fix attempt")
-        else:
-            logger.debug("  ✓ Fixed all invalid geometries")
-    else:
-        logger.debug("  ✓ All geometries are valid")
-
-    return gdf_optimized
-
-
-def _is_boolean_data(series: pd.Series) -> bool:
-    """Duck-type detection of boolean data."""
-    # Check if data looks like boolean
-    unique_vals = set(str(v).lower() for v in series.dropna().unique())
-    boolean_values = {"true", "false", "1", "0", "yes", "no"}
-    return len(unique_vals) <= 2 and unique_vals.issubset(boolean_values)
-
-
-def _is_count_field(col: str, series: pd.Series) -> bool:
-    """Detect count fields by pattern and data characteristics."""
-    # Pattern-based detection
-    if col.startswith("votes_") or col in ["TOTAL", "DEM", "REP", "NAV", "vote_margin"]:
-        return True
-
-    # Duck-type detection: integer data with reasonable range for vote counts
-    try:
-        numeric_data = pd.to_numeric(series.dropna(), errors="coerce")
-        if numeric_data.notna().any():
-            # Check if all values are non-negative integers (typical for counts)
-            is_integer = (numeric_data % 1 == 0).all()
-            is_non_negative = (numeric_data >= 0).all()
-            # Reasonable range for vote counts (0 to 50,000)
-            reasonable_range = (numeric_data <= 50000).all()
-            return bool(is_integer and is_non_negative and reasonable_range)
-    except Exception as e:
-        logger.debug(f"  ❌ Error detecting count field: {e}")
-
-    return False
-
-
-def _is_percentage_field(col: str, series: pd.Series) -> bool:
-    """Detect percentage fields by pattern and data characteristics."""
-    # Pattern-based detection
-    percentage_patterns = ["_pct_", "_rate", "_advantage", "_score", "_efficiency", "_potential"]
-    if any(pattern in col for pattern in percentage_patterns):
-        return True
-
-    # Duck-type detection: numeric data in percentage-like range
-    try:
-        numeric_data = pd.to_numeric(series.dropna(), errors="coerce")
-        if numeric_data.notna().any():
-            # Check if data looks like percentages (0-100 range mostly)
-            min_val, max_val = numeric_data.min(), numeric_data.max()
-            # Allow some flexibility for calculated fields that might go slightly outside 0-100
-            return bool(-200 <= min_val <= 200 and -200 <= max_val <= 200)
-    except Exception as e:
-        logger.debug(f"  ❌ Error detecting percentage field: {e}")
-
-    return False
-
-
-def _is_categorical_field(col: str, series: pd.Series) -> bool:
-    """Detect categorical fields by data characteristics."""
-    # Skip if looks like numeric data
-    try:
-        numeric_data = pd.to_numeric(series.dropna(), errors="coerce")
-        if numeric_data.notna().sum() > len(series.dropna()) * 0.8:  # 80% numeric
-            return False
-    except Exception as e:
-        logger.debug(f"  ❌ Error detecting categorical field: {e}")
-
-    # Check if it has limited unique values (typical for categories)
-    unique_count = series.nunique()
-    total_count = len(series.dropna())
-
-    # Consider categorical if: few unique values OR string-like data
-    if unique_count <= 20 or (total_count > 0 and unique_count / total_count < 0.1):
-        return True
-
-    # Check for common categorical indicators
-    sample_values = set(str(v).lower() for v in series.dropna().head(10))
-    categorical_indicators = {
-        "low",
-        "medium",
-        "high",
-        "small",
-        "large",
-        "strong",
-        "weak",
-        "competitive",
-        "safe",
-        "likely",
-        "tossup",
-        "close",
-        "clear",
-        "landslide",
-        "dem",
-        "rep",
-        "unknown",
-        "no data",
-        "tie",
-    }
-
-    return bool(len(sample_values & categorical_indicators) > 0)
-
-
-def _is_identifier_field(col: str) -> bool:
-    """Detect identifier/name fields by pattern."""
-    identifier_patterns = ["precinct", "candidate", "name", "id", "_id", "identifier"]
-    return any(pattern in col.lower() for pattern in identifier_patterns)
-
-
-def _optimize_boolean_field(series: pd.Series) -> pd.Series:
-    """Optimize boolean field for web display."""
-    return (
-        series.astype(str)
-        .str.lower()
-        .map({"true": True, "false": False, "1": True, "0": False, "yes": True, "no": False})
-        .fillna(False)
-    )
-
-
-def _optimize_count_field(series: pd.Series) -> pd.Series:
-    """Optimize count field for web display."""
-    numeric_series = pd.to_numeric(series, errors="coerce").fillna(0)
-    return numeric_series.astype(int)
-
-
-def _optimize_percentage_field(series: pd.Series, precision: int) -> pd.Series:
-    """Optimize percentage field for web display."""
-    numeric_series = pd.to_numeric(series, errors="coerce").fillna(0)
-    return numeric_series.round(precision)
-
-
-def _optimize_categorical_field(col: str, series: pd.Series) -> pd.Series:
-    """Optimize categorical field for web display."""
-    optimized = series.astype(str).replace(["nan", "None", "<NA>", ""], "No Data")
-
-    # Set appropriate defaults for specific field patterns
-    if "political" in col.lower() or "lean" in col.lower():
-        optimized = optimized.replace("No Data", "Unknown")
-    elif "competitive" in col.lower():
-        optimized = optimized.replace("No Data", "No Election Data")
-    elif "candidate" in col.lower():
-        optimized = optimized.replace("No Data", "No Data")
-    elif "density" in col.lower():
-        # Preserve categorical text values for voter_density_category etc.
-        pass  # Keep "No Data" as is
-
-    return optimized
-
-
-def _optimize_identifier_field(series: pd.Series) -> pd.Series:
-    """Optimize identifier field for web display."""
-    return series.astype(str).str.strip()
-
-
-def _optimize_unknown_field(col: str, series: pd.Series, precision: int) -> pd.Series:
-    """Fallback optimization for unknown field types."""
-    logger.debug(f"    ❓ Unknown field type for '{col}', attempting auto-detection")
-
-    # Try numeric first
-    try:
-        numeric_series = pd.to_numeric(series, errors="coerce")
-        if numeric_series.notna().sum() > len(series) * 0.7:  # 70% numeric
-            # If mostly integers, treat as count
-            if (numeric_series.dropna() % 1 == 0).all():
-                return numeric_series.fillna(0).astype(int)
-            # Otherwise, treat as decimal with precision
-            else:
-                return numeric_series.fillna(0).round(precision)
-    except Exception as e:
-        logger.debug(f"  ❌ Error optimizing unknown field: {e}")
-
-    # Fall back to string
-    return series.astype(str).str.strip()
 
 
 def create_candidate_color_mapping(candidate_cols: List[str]) -> Dict[str, str]:
